Consolidate config to pbnh/__init__.py
