--- pbnh/conf.py
@@ -1,27 +1,0 @@
-import yaml
-import copy
-
-DEFAULTS = {
-    "server": {
-        "bind_ip": "127.0.0.1",
-        "bind_port": 5001,
-        "debug": True,
-    },
-    "database": {
-        "dbname": "pastedb",
-        "dialect": "sqlite",
-        "driver": None,
-        "host": None,
-        "password": None,
-        "port": None,
-        "username": None,
-    },
-}
-
-
-def get_config():
-    try:
-        with open("secrets.yml") as config:
-            return yaml.safe_load(config)
-    except IOError:
-        return copy.copy(DEFAULTS)
